Serving/server/pyproject.toml
[tool.poetry]
name = "ai"
version = "0.1.0"
description = "Serving for AI services"
authors = ["harry <harry@opentechiz.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.5"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
pydantic = {extras = ["dotenv"], version = "^1.10.5"}
requests = "^2.28.2"
pillow = "^9.4.0"
facenet-pytorch = "^2.5.2"
torch = "1.7.1"
torchvision = "0.8.2"
pandas = "^1.5.3"
opencv-python-headless = "4.5.2.54"
PyYAML = "^6.0"
tqdm = "^4.65.0"
matplotlib = "^3.4.3"
seaborn = "^0.11.2"
psutil = "5.9.4"
aiohttp = "3.8.4"
grpcio = "1.54.0"
grpcio-tools = "1.54.0"
tensorflow-serving-api = "2.12.1"
tensorflow = "2.12.0"
ultralytics = "^8.0.105"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core==1.4.2"]
build-backend = "poetry.core.masonry.api"
